// src/providers/types/api-private.ts
export type ControlMode = '0_NO_CONTROL'|'1_CONTROL'|'2_SOB_DEMANDA'|'3_BACKUP'|'4_BLOCKED'|'5_BACKUP_CONTROL'|'6_BACKUP_CONTROL_V2'|'7_FORCED'|'8_ECO_2';
// ...
export interface ScheduleDut {
  DUT_SCHEDULE_ID?: number
  SCHEDULE_TITLE: string
  SCHEDULE_STATUS: boolean
  BEGIN_TIME: string
  END_TIME: string
  CTRLOPER: ControlMode
  PERMISSION: 'allow'|'forbid'
  DAYS: {
    mon: boolean
    tue: boolean
    wed: boolean
    thu: boolean
    fri: boolean
    sat: boolean
    sun: boolean
  }
  SETPOINT?: number|null
  LTC?: number|null
  LTI?: number|null
  UPPER_HYSTERESIS?: number|null
  LOWER_HYSTERESIS?: number|null
  SCHEDULE_START_BEHAVIOR?: string|null
  SCHEDULE_END_BEHAVIOR?: string|null
  FORCED_BEHAVIOR?: string|null
  IR_ID_COOL?: string|null
  ACTION_MODE?: string|null
  ACTION_TIME?: number|null
  ACTION_POST_BEHAVIOR?: string|null
}